pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
grpcio = "^1.71.0"
grpcio-tools = "^1.71.0"
yoyo-migrations = "^9.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
sqlalchemy = "^2.0.40"
requests = "^2.32.3"
python-keycloak = "^5.5.0"
